+++ src/LL/LLCref.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import type { LLText } from '../LLText/LLText';
+import findLatexCommandMatches from '../util/findLatexCommandMatches';
 import { messages } from '../util/constants';
-import match2range from '../util/match2range';
 import ranges2diagnostics from '../util/ranges2diagnostics';
 
 export default function LLCref(doc: vscode.TextDocument, txt: LLText, exceptions: string[]): vscode.Diagnostic[] {
@@ -11,13 +11,16 @@ export default function LLCref(doc: vscode.TextDocument, txt: LLText, exceptions
     const message: string[] = [];
     const ranges: vscode.Range[] = [];
 
-    for (const match of txt.text.matchAll(/\\ref\{([^}]*)/g)) {
+    for (const match of findLatexCommandMatches(txt.text, /\\ref\{/g)) {
         if (txt.isPreamble(match.index)) continue;
         if (!txt.isValid(match.index)) continue;
-        const refContent = match[1];
+        const refContent = match.content;
         if (exceptions.some(exception => refContent.startsWith(exception))) continue;
         message.push(messages[code]);
-        ranges.push(match2range(doc, match));
+        ranges.push(new vscode.Range(
+            doc.positionAt(match.index),
+            doc.positionAt(match.fullEnd)
+        ));
     }
 
     return ranges2diagnostics(code, message, ranges);