+++ src/LL/LLUnRef.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import type { LLText } from '../LLText/LLText';
+import findLatexCommandMatches from '../util/findLatexCommandMatches';
 import { messages } from '../util/constants';
-import match2range from '../util/match2range';
 import ranges2diagnostics from '../util/ranges2diagnostics';
 
 
@@ -16,10 +16,10 @@ export default function LLUnRef(doc: vscode.TextDocument, txt: LLText): vscode.D
         const foundRefs = new Set<string>();
         // This regex matches LaTeX commands that contain "ref" (case-insensitive).
         // If we lack the "ref" part, we might match other commands like \label.
-        const refRegex = /\\[a-zA-Z]*(?:ref|Ref|REF)[a-zA-Z]*\{([^}]*)\}/g;
-        for (const refMatch of txt.text.matchAll(refRegex)) {
-            if (!txt.isValid(refMatch.index!)) continue;
-            const refNames = refMatch[1].split(',').map(name => name.trim());
+        const refRegex = /\\[a-zA-Z]*(?:ref|Ref|REF)[a-zA-Z]*\{/g;
+        for (const refMatch of findLatexCommandMatches(txt.text, refRegex)) {
+            if (!txt.isValid(refMatch.index)) continue;
+            const refNames = refMatch.content.split(',').map(name => name.trim());
             for (const refName of refNames) foundRefs.add(refName);
         }
         return foundRefs;
@@ -51,17 +51,19 @@ export default function LLUnRef(doc: vscode.TextDocument, txt: LLText): vscode.D
             if (!endMatch) continue;
 
             const envContent = txt.text.substring(beginMatch.index + envHeaderLength, endMatch.index);
-            const labelRegex = /\\label\{([^}]*)\}/g;
+            const labelRegex = /\\label\{/g;
 
-            let labelMatch: RegExpExecArray | null;
-            while ((labelMatch = labelRegex.exec(envContent)) !== null) {
-                const labelName = labelMatch[1];
-                const absoluteIndex = beginMatch.index + envHeaderLength + labelMatch.index!;
+            for (const labelMatch of findLatexCommandMatches(envContent, labelRegex)) {
+                const labelName = labelMatch.content;
+                const absoluteIndex = beginMatch.index + envHeaderLength + labelMatch.index;
                 if (!txt.isValid(absoluteIndex)) continue;
                 if (allRefs.has(labelName)) continue;
 
                 message.push(messages[code].replaceAll("%1", labelName).replaceAll("%2", envName));
-                ranges.push(match2range(doc, { ...labelMatch, index: absoluteIndex }));
+                ranges.push(new vscode.Range(
+                    doc.positionAt(absoluteIndex),
+                    doc.positionAt(beginMatch.index + envHeaderLength + labelMatch.fullEnd)
+                ));
             }
         }
     };